pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yamyam-agent"
version = "0.1.0"
description = "Yamyam MCP Agent using Gemini API and LangChain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.1",
    "google-genai>=0.2.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.8",
    "mcp>=0.9.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
yamyam-agent = "yamyam_agent.cli:main"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["yamyam_agent"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["yamyam_agent"]
